+++ mpm/2D/Core/Element.hpp
@@ -1,8 +1,10 @@
 #ifndef ELEMENT_HPP
 #define ELEMENT_HPP
 
+#include <cstddef>
+
 struct element {
-  static int nbNodes;
+  static size_t nbNodes;
 
   //    13 12 11 10
   // 	  14 3  2  9
@@ -10,7 +12,7 @@ struct element {
   //    4  5  6  7
   // for regularquadlinear we use the inner square only
   // for BsplineCubic we use the inner and outer square
-  int I[16];
+  size_t I[16];
 
   element();  // Ctor
 };
